Put containers in a separate file Added warning page in case of no flatfield events found
